+++ crates/compiler/tests/fixtures/tree_shake/dynamic-import/output.js
@@ -13,7 +13,7 @@
     }
     require(entryModule);
 })({
-    "d2214aaa": function(module, exports, require, dynamicRequire) {
+    "d2214aaa": function(module, exports, farmRequire, dynamicRequire) {
         "use strict";
         console.log("runtime/index.js");
         __farm_global_this__.__farm_module_system__.setPlugins([]);
@@ -25,7 +25,7 @@
         __farm_global_this__.__farm_module_system__.register(key, modules[key]);
     }
 })({
-    "b5d64806": function(module, exports, require, dynamicRequire) {
+    "b5d64806": function(module, exports, farmRequire, dynamicRequire) {
         "use strict";
         dynamicRequire("05ee5ec7").then((dep)=>{
             console.log(dep);
@@ -45,7 +45,7 @@ var entry = farmModuleSystem.require("b5d64806");
         __farm_global_this__.__farm_module_system__.register(key, modules[key]);
     }
 })({
-    "05ee5ec7": function(module, exports, require, dynamicRequire) {
+    "05ee5ec7": function(module, exports, farmRequire, dynamicRequire) {
         "use strict";
         Object.defineProperty(exports, "__esModule", {
             value: true
@@ -64,15 +64,15 @@ var entry = farmModuleSystem.require("b5d64806");
                 return _default;
             }
         });
-        var _interop_require_default = require("@swc/helpers/_/_interop_require_default");
-        var _dep1 = _interop_require_default._(require("ef0c4c9d"));
+        var _interop_require_default = farmRequire("@swc/helpers/_/_interop_require_default");
+        var _dep1 = _interop_require_default._(farmRequire("ef0c4c9d"));
         const dep = "dep";
         function _default() {
             return (0, _dep1.default)();
         }
         console.log("side effect in dep.ts");
     },
-    "ef0c4c9d": function(module, exports, require, dynamicRequire) {
+    "ef0c4c9d": function(module, exports, farmRequire, dynamicRequire) {
         "use strict";
         Object.defineProperty(exports, "__esModule", {
             value: true