+++ crates/plugin_script/src/lib.rs
@@ -5,7 +5,7 @@ use std::sync::Arc;
 
 use deps_analyzer::DepsAnalyzer;
 use farmfe_core::{
-  config::{Config, TargetEnv, FARM_GLOBAL_THIS, FARM_MODULE_SYSTEM},
+  config::{Config, ModuleFormat, TargetEnv, FARM_GLOBAL_THIS, FARM_MODULE_SYSTEM},
   context::CompilationContext,
   error::{CompilationError, Result},
   module::{ModuleId, ModuleMetaData, ModuleSystem, ModuleType, ScriptModuleMetaData},
@@ -15,10 +15,10 @@ use farmfe_core::{
   },
   relative_path::RelativePath,
   resource::{
-    resource_pot::{self, ResourcePot, ResourcePotType},
+    resource_pot::{ResourcePot, ResourcePotType},
     Resource, ResourceType,
   },
-  swc_common::{comments::NoopComments, FilePathMapping, Mark, SourceMap, GLOBALS},
+  swc_common::{comments::NoopComments, Mark, GLOBALS},
   swc_ecma_ast::{
     CallExpr, Callee, Expr, ExprStmt, Ident, MemberExpr, MemberProp, ModuleDecl, ModuleItem, Stmt,
   },
@@ -31,20 +31,17 @@ use farmfe_toolkit::{
     swc_try_with::try_with, syntax_from_module_type,
   },
   sourcemap::swc_gen::{build_source_map, AstModule},
-  swc_ecma_minifier::{
-    optimize,
-    option::{ExtraOptions, MinifyOptions},
-  },
   swc_ecma_transforms::{
-    fixer::fixer,
     resolver,
     typescript::{strip, strip_with_jsx},
   },
-  swc_ecma_visit::FoldWith,
   swc_ecma_visit::VisitMutWith,
 };
+use swc_plugins::transform_by_swc_plugins;
 
 mod deps_analyzer;
+mod swc_plugins;
+
 /// ScriptPlugin is used to support compiling js/ts/jsx/tsx/... files, support loading, parse, analyze dependencies and code generation.
 /// Note that we do not do transforms here, the transforms (e.g. strip types, jsx...) are handled in a separate plugin (farmfe_plugin_swc_transforms).
 pub struct FarmPluginScript {}
@@ -152,6 +149,7 @@ impl Plugin for FarmPluginScript {
             farmfe_core::module::ModuleType::Tsx => {
               ast.visit_mut_with(&mut strip_with_jsx(
                 context.meta.script.cm.clone(),
+                // TODO make it configurable
                 Default::default(),
                 NoopComments, // TODO parse comments
                 top_level_mark,
@@ -161,11 +159,18 @@ impl Plugin for FarmPluginScript {
           }
         },
       )?;
-
-      return Ok(Some(()));
     }
 
-    Ok(None)
+    // execute swc plugins
+    if param.module_type.is_script() && context.config.script.plugins.len() > 0 {
+      try_with(
+        context.meta.script.cm.clone(),
+        &context.meta.script.globals,
+        || transform_by_swc_plugins(param, context).unwrap(),
+      )?;
+    }
+
+    Ok(Some(()))
   }
 
   fn analyze_deps(
@@ -179,8 +184,10 @@ impl Plugin for FarmPluginScript {
       let module_ast = &module.meta.as_script().ast;
       // TODO deal with dynamic import, when dynamic import and static import are mixed, using static import
       let mut analyzer = DepsAnalyzer::new(
+        &module.id,
         module_ast,
         Mark::from_u32(module.meta.as_script().unresolved_mark),
+        Mark::from_u32(module.meta.as_script().top_level_mark),
       );
 
       GLOBALS.set(&context.meta.script.globals, || {
@@ -203,23 +210,28 @@ impl Plugin for FarmPluginScript {
     if !param.module.module_type.is_script() {
       return Ok(None);
     }
-
+    // all jsx, js, ts, tsx modules should be transformed to js module for now
+    // cause the partial bundling is not support other module type yet
     param.module.module_type = ModuleType::Js;
-
-    if param.deps.len() > 0 {
-      let module_system =
-        module_system_from_deps(param.deps.iter().map(|d| d.kind.clone()).collect());
-      param.module.meta.as_script_mut().module_system = module_system;
+    // default to commonjs
+    let module_system = if param.deps.len() > 0 {
+      module_system_from_deps(param.deps.iter().map(|d| d.kind.clone()).collect())
     } else {
-      // default to commonjs
-      param.module.meta.as_script_mut().module_system = ModuleSystem::CommonJs;
+      ModuleSystem::CommonJs
+    };
+    param.module.meta.as_script_mut().module_system = module_system.clone();
 
-      let ast = &param.module.meta.as_script().ast;
+    let ast = &param.module.meta.as_script().ast;
 
+    if module_system != ModuleSystem::Hybrid {
       // if the ast contains ModuleDecl, it's a esm module
       for item in ast.body.iter() {
         if let ModuleItem::ModuleDecl(_) = item {
-          param.module.meta.as_script_mut().module_system = ModuleSystem::EsModule;
+          if module_system == ModuleSystem::CommonJs && param.deps.len() > 0 {
+            param.module.meta.as_script_mut().module_system = ModuleSystem::Hybrid;
+          } else {
+            param.module.meta.as_script_mut().module_system = ModuleSystem::EsModule;
+          }
           break;
         }
       }
@@ -436,14 +448,19 @@ impl FarmPluginScript {
             .insert(0, runtime_ast.body.to_vec().remove(0));
 
           let export_info = self.get_export_info_of_entry_module(entry_module_id, context);
-          let export_str = if context.config.output.target_env == TargetEnv::Node {
+          let export_str = if export_info.len() > 0 {
             export_info
               .iter()
               .map(|export| {
                 if export == "default" {
-                  "export default entry.default;".to_string()
+                  match context.config.output.format {
+                    ModuleFormat::CommonJs => "module.exports = entry.default;".to_string(),
+                    ModuleFormat::EsModule => "export default entry.default;".to_string(),
+                    _ => panic!("default export is not supported in this format"),
+                  }
                 } else {
-                  format!("export {{ {}: entry.{} }};", export, export)
+                  // format!("export {{ {}: entry.{} }};", export, export)
+                  panic!("named export is not supported");
                 }
               })
               .collect::<Vec<String>>()
@@ -475,11 +492,21 @@ impl FarmPluginScript {
           // insert node specific code.
           // TODO: support async module for node, using dynamic require to load external module instead of createRequire. createRequire does not support load ESM module.
           if context.config.output.target_env == TargetEnv::Node {
+            let code_str = match context.config.output.format {
+              ModuleFormat::EsModule => {
+                r#"import module from 'node:module';
+                global.__farmNodeRequire = module.createRequire(import.meta.url);
+                global.__farmNodeBuiltinModules = module.builtinModules;"#
+              }
+              ModuleFormat::CommonJs => {
+                r#"global.__farmNodeRequire = require;
+                global.__farmNodeBuiltinModules = require('node:module').builtinModules;"#
+              }
+              _ => panic!("node only support cjs and esm format"),
+            };
             let global_var = parse_module(
               "farm-global-var",
-              r#"import module from 'node:module';
-                global.__farmNodeRequire = module.createRequire(import.meta.url);
-                global.__farmNodeBuiltinModules = module.builtinModules;"#,
+              code_str,
               Syntax::Es(context.config.script.parser.es_config.clone()),
               context.config.script.target.clone(),
               context.meta.script.cm.clone(),