+++ examples/vue-antdv/src/test1.vue
@@ -1,0 +1,13 @@
+<template>
+  <div>
+    test1
+    <RouterView />
+  </div>
+</template>
+
+<script lang="ts" setup>
+import { ref } from 'vue';
+import { DatePicker } from 'ant-design-vue';
+
+const selectedDate = ref<string>();
+</script>