+++ packages/core/binding/index.d.ts
@@ -110,6 +110,7 @@ export interface Config {
       publicPath?: string;
       assetsFilename?: string;
       targetEnv?: 'browser' | 'node';
+      format?: 'cjs' | 'esm';
     };
     resolve?: {
       extensions?: string[];
@@ -172,6 +173,14 @@ export interface Config {
           noEarlyErrors: boolean;
         };
       };
+      plugins: {
+        name: string;
+        options?: Record<string, any>;
+        filters?: {
+          resolvedPaths?: string[];
+          moduleTypes?: ModuleType[];
+        }
+      }[];
     };
     css?: {
       modules?: {