+++ packages/core/src/config/schema.ts
@@ -12,7 +12,8 @@ const ConfigSchema = z
         path: z.string().optional(),
         publicPath: z.string().optional(),
         assetsFilename: z.string().optional(),
-        targetEnv: z.enum(['browser', 'node']).optional()
+        targetEnv: z.enum(['browser', 'node']).optional(),
+        format: z.enum(['cjs', 'esm']).optional()
       })
       .strict()
       .optional(),
@@ -61,31 +62,34 @@ const ConfigSchema = z
             'es2022'
           ])
           .optional(),
-        parser: z.object({
-          esConfig: z
-            .object({
-              jsx: z.boolean().optional(),
-              fnBind: z.boolean(),
-              decorators: z.boolean(),
-              decoratorsBeforeExport: z.boolean(),
-              exportDefaultFrom: z.boolean(),
-              importAssertions: z.boolean(),
-              privateInObject: z.boolean(),
-              allowSuperOutsideMethod: z.boolean(),
-              allowReturnOutsideFunction: z.boolean()
-            })
-            .strict()
-            .optional(),
-          tsConfig: z
-            .object({
-              tsx: z.boolean(),
-              decorators: z.boolean(),
-              dts: z.boolean(),
-              noEarlyErrors: z.boolean()
-            })
-            .strict()
-            .optional()
-        })
+        parser: z
+          .object({
+            esConfig: z
+              .object({
+                jsx: z.boolean().optional(),
+                fnBind: z.boolean(),
+                decorators: z.boolean(),
+                decoratorsBeforeExport: z.boolean(),
+                exportDefaultFrom: z.boolean(),
+                importAssertions: z.boolean(),
+                privateInObject: z.boolean(),
+                allowSuperOutsideMethod: z.boolean(),
+                allowReturnOutsideFunction: z.boolean()
+              })
+              .strict()
+              .optional(),
+            tsConfig: z
+              .object({
+                tsx: z.boolean(),
+                decorators: z.boolean(),
+                dts: z.boolean(),
+                noEarlyErrors: z.boolean()
+              })
+              .strict()
+              .optional()
+          })
+          .optional(),
+        plugins: z.array(z.any()).optional()
       })
       .strict()
       .optional(),