+++ examples/script-entry/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @farmfe-examples/script-entry
 
+## 0.0.24
+
+### Patch Changes
+
+- Updated dependencies [e0ff7f1]
+  - @farmfe/core@0.8.3
+  - @farmfe/cli@0.5.2
+
+## 0.0.23
+
+### Patch Changes
+
+- Updated dependencies [1148f68]
+  - @farmfe/core@0.8.2
+  - @farmfe/cli@0.5.2
+
+## 0.0.22
+
+### Patch Changes
+
+- Updated dependencies [6f97c87]
+  - @farmfe/core@0.8.1
+  - @farmfe/cli@0.5.2
+
+## 0.0.21
+
+### Patch Changes
+
+- Updated dependencies [e780747]
+  - @farmfe/core@0.8.0
+  - @farmfe/cli@0.5.2
+
 ## 0.0.20
 
 ### Patch Changes