+++ js-plugins/less/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @farmfe/js-plugin-less
 
+## 0.0.4
+
+### Patch Changes
+
+- 1148f68: Bugfix https://github.com/farm-fe/farm/issues/336
+- Updated dependencies [1148f68]
+  - @farmfe/core@0.8.2
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [e780747]
+  - @farmfe/core@0.8.0
+
 ## 0.0.2
 
 ### Patch Changes